[ROCm] AllReduce+RMSNorm Fusion type fix
In cases like Qwen3Next, the weight tensor coming in is a float32 or non
bf16 tensor. The aiter kernel used for fusin all_reduce+rms_norm expects
the opaque pointer given to it to be the same as the input type.

+++ vllm/compilation/passes/fusion/allreduce_rms_fusion.py
@@ -943,7 +943,7 @@ class AiterAllreduceFusedRMSNormPattern(BasePattern, VllmPatternReplacement):
             allreduce = self.FUSED_AR_RMSNORM_OP(
                 input_=input,
                 residual=residual,
-                weight=weight,
+                weight=weight.to(input.dtype),
                 epsilon=self.epsilon,
             )
             return allreduce[0], allreduce[1]
@@ -989,7 +989,7 @@ class AiterAllreduceFusedAddRMSNormPattern(BasePattern, VllmPatternReplacement):
             allreduce = self.FUSED_AR_RMSNORM_OP(
                 input_=input,
                 residual=residual,
-                weight=weight,
+                weight=weight.to(input.dtype),
                 epsilon=self.epsilon,
             )
             return allreduce[0], allreduce[1]
